update replica movement acceptance test
