+++ test/acceptance/replication/replica_replication/fast/delete_class_test.go
@@ -35,6 +35,7 @@ func (suite *ReplicationTestSuiteEndpoints) TestReplicationDeletingClassCleansUp
 
 	compose, err := docker.New().
 		WithWeaviateCluster(3).
+		WithWeaviateEnv("REPLICA_MOVEMENT_MINIMUM_ASYNC_WAIT", "5s").
 		Start(mainCtx)
 	require.Nil(t, err)
 	defer func() {
