+++ test/acceptance/replication/replica_replication/fast/delete_tenant_test.go
@@ -37,6 +37,7 @@ func (suite *ReplicationTestSuiteEndpoints) TestReplicationDeletingTenantCleansU
 
 	compose, err := docker.New().
 		WithWeaviateCluster(3).
+		WithWeaviateEnv("REPLICA_MOVEMENT_MINIMUM_ASYNC_WAIT", "5s").
 		Start(mainCtx)
 	require.Nil(t, err)
 	defer func() {
