+++ test/acceptance/replication/replica_replication/fast/endpoints_test.go
@@ -49,6 +49,7 @@ func (suite *ReplicationTestSuiteEndpoints) TestReplicationReplicateEndpoints() 
 
 	compose, err := docker.New().
 		WithWeaviateCluster(3).
+		WithWeaviateEnv("REPLICA_MOVEMENT_MINIMUM_ASYNC_WAIT", "5s").
 		Start(mainCtx)
 	require.Nil(t, err)
 	defer func() {
